+++ tests/test_arm.c
@@ -18,7 +18,7 @@ struct platform {
 	int syntax;
 };
 
-static void print_string_hex(char *comment, unsigned char *str, int len)
+static void print_string_hex(char *comment, unsigned char *str, size_t len)
 {
 	unsigned char *c;
